requirements.txt
streamlit
pymongo
pytesseract
opencv-python
pillow